// amplify/function/VerifyUserCode/graphql/API.ts
export type FoodEvent = {
  __typename: "FoodEvent",
  attended?: ModelUserConnection | null,
  createdAt: string,
  description?: string | null,
  end?: string | null,
  groups?: number | null,
  id: string,
  name?: string | null,
  owner?: string | null,
  start?: string | null,
  updatedAt: string,
};

export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};

export type User = {
  __typename: "User",
  allergies?: string | null,
  checkedIn?: boolean | null,
  createdAt: string,
  email?: string | null,
  firstName?: string | null,
  id: string,
  institution?: string | null,
  lastName?: string | null,
  meal?: FoodEvent | null,
  mealId?: string | null,
  meals?: boolean | null,
  owner?: string | null,
  team?: Team | null,
  teamId?: string | null,
  updatedAt: string,
};

export type Team = {
  __typename: "Team",
  createdAt: string,
  id: string,
  members?: ModelUserConnection | null,
  name?: string | null,
  owner?: string | null,
  updatedAt: string,
};